# Significantly improve performance by not using broadcasting for very small arrays