--- a/src/GraphType.jl
+++ b/src/GraphType.jl
@@ -148,12 +148,12 @@ mutable struct GraphData
         # equivalence classes (at the beginning each state represents just itself)
         eq_vn(v0, p0) = (v0 == spp[p0] ? nothing : pvers[p0][v0])
 
-        # Hot code
+        # Hot code, measure performance before changing
         eq_classes = Dict{UUID,Dict{InstState,Set{InstState}}}()
         for p0 = 1:np
             d = Dict{InstState, Set{InstState}}()
             for v0 = 1:spp[p0]
-                let p0 = p0
+                let p0 = p0 # Due to https://github.com/JuliaLang/julia/issues/15276
                     d[eq_vn(v0,p0)] = Set([eq_vn(v0,p0)])
                 end
             end
@@ -161,7 +161,7 @@ mutable struct GraphData
         end
 
         # the resolution log is actually initialized below
-        rlog = ResolveLog(uuid_to_name, false)
+        rlog = ResolveLog(uuid_to_name, verbose)
 
         data = new(pkgs, np, spp, pdict, pvers, vdict, uuid_to_name, pruned, eq_classes, rlog)
 
@@ -253,12 +253,14 @@ mutable struct Graph
         extra_uuids = union(keys(reqs), keys(fixed), map(fx->keys(fx.requires), values(fixed))...)
         extra_uuids ⊆ keys(versions) || error("unknown UUID found in reqs/fixed") # TODO?
 
+        # Type assert below due to https://github.com/JuliaLang/julia/issues/25918
         data = GraphData(versions, deps, compat, uuid_to_name, verbose = verbose)::GraphData
         pkgs, np, spp, pdict, pvers, vdict, rlog = data.pkgs, data.np, data.spp, data.pdict, data.pvers, data.vdict, data.rlog
 
         local extended_deps
-        let spp = spp
-            extended_deps = [Vector{Dict{Int,BitVector}}(spp[p0]-1) for p0 = 1:np]::Vector{Vector{Dict{Int, BitVector}}}
+        let spp = spp # Due to https://github.com/JuliaLang/julia/issues/15276
+            # Type assert below to help inference
+            extended_deps = [Vector{Dict{Int,BitVector}}(spp[p0]-1) for p0 = 1:np]::Vector{Vector{Dict{Int,BitVector}}}
         end
         for p0 = 1:np, v0 = 1:(spp[p0]-1)
             n2u = Dict{String,UUID}()
@@ -289,7 +291,7 @@ mutable struct Graph
                 get!(req, pdict[uuid]) do; VersionSpec() end
             end
             # Translate the requirements into bit masks
-            # Hot code
+            # Hot code, measure performance before changing
             req_msk = Dict{Int,BitVector}()
             @inbounds for (p1, vs) in req
                 pv = pvers[p1]
@@ -305,7 +307,7 @@ mutable struct Graph
         gadj = [Int[] for p0 = 1:np]
         gmsk = [BitMatrix[] for p0 = 1:np]
         local gconstr
-        let spp = spp
+        let spp = spp # Due to https://github.com/JuliaLang/julia/issues/15276
             gconstr = [trues(spp[p0]) for p0 = 1:np]
         end
         adjdict = [Dict{Int,Int}() for p0 = 1:np]