From profiling, a large bottleneck was found in creating the `Graph`. This is because broadcasting (while convenient) actually has a significant overhead when used on small arrays. This was frequently the case here.

Before the changes here, creating the `Graph` took 0.93 seconds when adding DifferentialEquations with a total time for `Pkg3.add` of 2.0 seconds.

After the changes, `Graph`takes 0.30 seconds and total time 1.36 seconds.

The changes to Operations.jl are for readability, was hard to me to keep the nested generators in my head (also this way we only do one loop).