@KristofferC 2018-02-07
From more profiling I added some more optimizations. These typically fixes manifestations of https://github.com/JuliaLang/julia/issues/15276 and some cases where generator are slow. This, unfortunately, makes the code a bit uglier but I think it is worth it.

With this PR creating the `Graph` takes:

```
  0.104140 seconds (545.37 k allocations: 35.921 MiB, 22.12% gc time)
```

compared to before

```
  1.026278 seconds (2.72 M allocations: 168.556 MiB, 19.84% gc time)
```

when adding DifferentialEquations.

With #116, we are now at a total time of 0.46 seconds to `Pkg3.add("DifferentialEquations")` compared to ~3seconds two days ago.