@carlobaldassi 2018-02-07
I also vote for performance over code aesthetics, the difference seems quite significant.