@KristofferC 2018-02-07
Added some comments explaining the reasons for some type asserts and `let`. Fixed the spurious space.